requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
pymysql-0.10.1
pyjwt-1.7.1
passlib-1.7.4
enum34-1.1.10